v2.3.7 - Fixes

+++ src/net/vadamdev/viaapi/tools/images/ImageDecoder.java
@@ -5,6 +5,7 @@ import net.vadamdev.viaapi.tools.images.particle.PixelParticle;
 import net.vadamdev.viaapi.tools.utils.FileUtil;
 
 import javax.xml.parsers.SAXParser;
+import java.awt.*;
 import java.awt.image.BufferedImage;
 
 public class ImageDecoder {
@@ -19,11 +20,11 @@ public class ImageDecoder {
         PixelParticle[] pixelParticles = new PixelParticle[image.getWidth() * image.getHeight()];
 
         int i = 0;
-        for(int x = 0; x < image.getWidth(); x++) {
-            for(int y = 0; y < image.getHeight(); y++) {
-                if(image.getRGB(x, y) == 0) continue;
+        for(double x = 0; x < image.getWidth(); x++) {
+            for(double y = 0; y < image.getHeight(); y++) {
+                if(image.getRGB((int) x, (int) y) == 0) continue;
 
-                pixelParticles[i] = new PixelParticle(x / PARTICLE_SPACE, y / PARTICLE_SPACE, image.getRGB(x, y));
+                pixelParticles[i] = new PixelParticle(x / PARTICLE_SPACE, y / PARTICLE_SPACE, image.getRGB((int) x, (int) y));
                 i++;
             }
         }
